% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_nbs.R
\name{cmp.nbs}
\alias{cmp.nbs}
\title{Compare neighbouring pixel values of a raster object}
\usage{
cmp.nbs(x, mode = "rook", assign = NA)
}
\arguments{
\item{x}{Raster data input.}

\item{mode}{Number of adjacent pixels to include in the comparison. Either a numeric value (4, 8 or 16) or a character string ("rook", "queen" or "bishop).}

\item{assign}{Value to assign to a pixel with different value to at least one of its neighbouring pixels.}
}
\value{
A raster with the dimensions of the original raster and either numeric values or NA.
}
\description{
Compares each pixel of a raster object with its adjacent pixels and replaces differing values by a user defined value.
}
